requirements.txt
flask
openai
python-dotenv
requests
google-generativeai
